Move models to one file
