+++ walledeval/judge/huggingface/__init__.py
@@ -1,4 +1,2 @@
 from walledeval.judge.huggingface.core import HFTextClassificationJudge
-
-from walledeval.judge.huggingface.gptfuzz import GPTFuzzJudge
-from walledeval.judge.huggingface.toxicbert import ToxicBertJudge
+from walledeval.judge.huggingface.models import GPTFuzzJudge, UnitaryJudge, RobertaToxicityJudge
