--- walledeval/judge/huggingface/gptfuzz.py
@@ -1,6 +1,0 @@
-from walledeval.judge.huggingface.core import HFTextClassificationJudge
-class GPTFuzzJudge(HFTextClassificationJudge[int, bool]):
-    def __init__(self):
-        super().__init__('hubert233/GPTFuzz')
-    def score(self, intermediate: int) -> bool:
-            return intermediate == 0
